move tests to test_fetchers.py
